feat: ad new module docs_consumer to get mosqlimate docs
