+++ mosqlimate_assistant/__init__.py
@@ -5,6 +5,7 @@ from . import (
     api_consumer,
     assistant,
     assistant_ollama,
+    docs_consumer,
     faiss_db,
     muni_codes,
     schemas,
@@ -24,14 +25,15 @@ def get_version() -> str:
 version: str = get_version()
 __version__: str = version
 __all__: List[str] = [
-    "assistant",
-    "assistant_ollama",
-    "settings",
-    "schemas",
-    "faiss_db",
-    "utils",
     "api_consumer",
+    "assistant_ollama",
+    "assistant",
+    "docs_consumer",
+    "faiss_db",
     "muni_codes",
+    "schemas",
+    "settings",
+    "utils",
     "eng",
     "por",
 ]  # noqa: WPS410 (the only __variable__ we use)
